Add XLineFollowerBuilder, implement most of XLineFollower, and fix major bugs
Add XLineFollowerBuilder so that you can configure it to a much more
advanced level.

Implement most of XLineFollower so that it can follow a straight line
(albeit pretty slowly)

Add a LineSensor class so that line sensors can have their factory error
fixed.

Have AutoChassis be able to clone itself very cleanly.

Bugs identified and fixed:
* AutoXChassis would crash if you did not call TurnA
* Line sensors would not settle to the similar value when set on the
  same surface, showing a constant error between their values.
* Misc. fixes

+++ src/ControllerCallback.cpp
@@ -31,7 +31,7 @@ void ControllerCallback::SyncCallback(controller_digital_e_t button, task_fn_t f
 
 
 void ControllerCallback::PollController() {
-	for(int i = 0; i < DIGITAL.size(); i++) {
+	for(size_t i = 0; i < DIGITAL.size(); i++) {
 		controller_digital_e_t button = DIGITAL[i];
 
 		if(callbackButtonMap.find(button) != callbackButtonMap.end()) {
